requirements.txt
flask
flask-cors
networkx
pydriller
groq
astunparse
python-dotenv
